packages/tools/skill-2-scenario-factory/requirements.txt
google-genai>=1.0.0
supabase>=2.0.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
httpx>=0.27.0
